fix(storage-pg): bind/decode latency_ms as bigint in the persist verb + test
The persist_mcp_call verb has its own direct mcp_call_logged_v1 INSERT (separate from the macro
sidecar); after widening latency_ms to bigint it must bind i64::from(u32) (lossless, removes the
i32::MAX cap) and the integration test must decode latency_ms as i64. PG-only fallout from the
column widen (cdx is blind to PG).

+++ crates/storage-pg/tests/integration/persist_mcp_call.rs
@@ -28,7 +28,7 @@ async fn persist_writes_fact_inline_io_citation_and_change_event() {
                 .await?;
         assert_eq!(n_facts.0, 1);
 
-        let fact: (String, String, bool, i32) = sqlx::query_as(
+        let fact: (String, String, bool, i64) = sqlx::query_as(
             "SELECT tool_name, actor_upn, ok, latency_ms \
              FROM proxima_core.mcp_call_logged_v1 WHERE memory_id = $1",
         )
